+++ webpack.config.js
@@ -1,80 +1,13 @@
 'use strict';
 
 const path = require('path');
-const webpack = require('webpack');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 const proxy = require('./server/webpack-dev-proxy');
-const StyleLintPlugin = require('stylelint-webpack-plugin');
-const SplitByPathPlugin = require('webpack-split-by-path');
-
 const loaders = require('./webpack/loaders');
-
-const basePlugins = [
-  new webpack.DefinePlugin({
-    __DEV__: process.env.NODE_ENV !== 'production',
-    __PRODUCTION__: process.env.NODE_ENV === 'production',
-    'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-  }),
-  new SplitByPathPlugin([
-    { name: 'vendor', path: [__dirname + '/node_modules/'] },
-  ]),
-  new HtmlWebpackPlugin({
-    template: './src/index.html',
-    inject: 'body',
-    minify: false,
-  }),
-  new webpack.NoErrorsPlugin(),
-];
-
-const devPlugins = [
-  new StyleLintPlugin({
-    configFile: './.stylelintrc',
-    files: 'src/**/*.css',
-    failOnError: false,
-  }),
-];
-
-const prodPlugins = [
-  new webpack.optimize.OccurenceOrderPlugin(),
-  new webpack.optimize.DedupePlugin(),
-  new webpack.optimize.UglifyJsPlugin({
-    mangle: true,
-    compress: {
-      warnings: false,
-    },
-  }),
-];
-
-const plugins = basePlugins
-  .concat(process.env.NODE_ENV === 'production' ? prodPlugins : [])
-  .concat(process.env.NODE_ENV === 'development' ? devPlugins : []);
-
-const postcssBasePlugins = [
-  require('postcss-import')({
-    addDependencyTo: webpack,
-  }),
-  require('postcss-cssnext')({
-    browsers: ['ie >= 8', 'last 2 versions'],
-  }),
-];
-const postcssDevPlugins = [];
-const postcssProdPlugins = [
-  require('cssnano')({
-    safe: true,
-    sourcemap: true,
-    autoprefixer: false,
-  }),
-];
-
-const postcssPlugins = postcssBasePlugins
-  .concat(process.env.NODE_ENV === 'production' ? postcssProdPlugins : [])
-  .concat(process.env.NODE_ENV === 'development' ? postcssDevPlugins : []);
+const plugins = require('./webpack/plugins');
+const postcssInit = require('./webpack/postcss');
 
 module.exports = {
-  entry: {
-    app: './src/index.ts',
-  },
-
+  entry: { app: './src/index.ts' },
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: '[name].[hash].js',
@@ -87,11 +20,9 @@ module.exports = {
     'source-map' :
     'inline-source-map',
 
-  resolve: {
-    extensions: ['', '.webpack.js', '.web.js', '.ts', '.js'],
-  },
-
+  resolve: { extensions: ['', '.webpack.js', '.web.js', '.ts', '.js'] },
   plugins: plugins,
+  postcss: postcssInit,
 
   devServer: {
     historyApiFallback: { index: '/' },
@@ -113,9 +44,5 @@ module.exports = {
       loaders.ttf,
     ],
     noParse: [ /zone\.js\/dist\/.+/, /angular2\/bundles\/.+/ ],
-  },
-
-  postcss: function postcssInit() {
-    return postcssPlugins;
   },
 };
